pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["lcpires7 <lcpires7.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
djangorestframework = "^3.16.0"
django-extensions = "^4.1"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
